Rename `getServicesState` -> `getServiceTargetStates`
