+++ server/modules/environment-state/getASGState.js
@@ -4,7 +4,7 @@
 let _ = require('lodash');
 let co = require('co');
 let getInstanceState = require('./getInstanceState');
-let getServicesState = require('./getServicesState');
+let getServiceTargetStates = require('./getServiceTargetStates');
 let AutoScalingGroup = require('models/AutoScalingGroup');
 let resourceProvider = require('modules/resourceProvider');
 let logger = require('modules/logger');
@@ -62,7 +62,7 @@ module.exports = function getASGState(accountName, environmentName, asgName) {
 
     let response = {
       Instances: instances,
-      Services: yield getServicesState(environmentName, asg.getRuntimeServerRoleName(), instances)
+      Services: yield getServiceTargetStates(environmentName, asg.getRuntimeServerRoleName(), instances)
     };
     return response;
   });
